refactor(dev): Move the live reload wiring into the dummy app
Disabling the gem's watcher only ever applies when Alchemy is developed from a
checkout, yet it shipped in the engine and ran for every application that
installs Alchemy, guarded by a root check that made it a no-op everywhere else.
The dummy app is the only place the situation arises, so the wiring belongs
there and the engine keeps no development-only initializers.

It has to live in application.rb rather than config/initializers. The gem
offers no hook to skip its watcher, so it is disabled only for the window
between the gem's middleware and watcher initializers: disabling any earlier
would also skip the middleware that injects the client script. Files in
config/initializers load while that chain is already being sorted, so they can
run code at a fixed point but cannot place an initializer inside the window.

Dropping the root check also drops the four examples covering it, since the
dummy app always sits inside the engine root. What remains is asserted against
Rails.application, including that the pair still resolves into the window.
