+++ spec/dummy/config/application.rb
@@ -36,5 +36,40 @@ module Dummy
     if ENV["ALCHEMY_STORAGE_ADAPTER"] == "active_storage"
       config.active_storage.variant_processor = :vips
     end
+
+    # This app lives inside the engine root, so rails_live_reload's own watcher
+    # and a watcher rooted at the engine cover the same tree and fight over the
+    # same socket. The gem offers no hook to skip its watcher, so it is disabled
+    # for the length of that single initializer and restored right after.
+    #
+    # The window matters: disabling any earlier would also skip the middleware
+    # that injects the client script, and config/initializers runs too early to
+    # define an initializer at all.
+    initializer "dummy.disable_live_reload_watcher",
+      after: "rails_live_reload.middleware",
+      before: "rails_live_reload.watcher" do
+      if Rails.env.development?
+        @live_reload_enabled = RailsLiveReload.enabled?
+        RailsLiveReload.config.enabled = false
+      end
+    end
+
+    initializer "dummy.restore_live_reload",
+      after: "rails_live_reload.watcher",
+      before: "rails_live_reload.configure_metrics" do
+      if Rails.env.development?
+        RailsLiveReload.config.enabled = @live_reload_enabled
+      end
+    end
+
+    # Started late so the watch patterns from the engine's initializer are in
+    # place, and so it replaces rather than joins the watcher disabled above.
+    config.after_initialize do
+      if Rails.env.development? && RailsLiveReload.enabled?
+        require "alchemy/dev_support/live_reload_watcher"
+
+        Alchemy::LiveReloadWatcher.init
+      end
+    end
   end
 end
