+++ spec/libraries/alchemy/engine_spec.rb
@@ -34,107 +34,6 @@ RSpec.describe Alchemy::Engine do
     end
   end
 
-  describe ".watch_engine_files?" do
-    subject { described_class.watch_engine_files? }
-
-    context "in development, with the application inside the engine root" do
-      before do
-        allow(Rails).to receive(:env).and_return(ActiveSupport::StringInquirer.new("development"))
-        allow(Rails.application).to receive(:root).and_return(described_class.root.join("spec/dummy"))
-      end
-
-      it { is_expected.to be(true) }
-    end
-
-    context "in development, with the application outside the engine root" do
-      before do
-        allow(Rails).to receive(:env).and_return(ActiveSupport::StringInquirer.new("development"))
-        allow(Rails.application).to receive(:root).and_return(Pathname.new("/somewhere/else"))
-      end
-
-      it { is_expected.to be(false) }
-    end
-
-    context "outside of development" do
-      before do
-        allow(Rails.application).to receive(:root).and_return(described_class.root.join("spec/dummy"))
-      end
-
-      it { is_expected.to be(false) }
-    end
-
-    context "without rails_live_reload" do
-      before do
-        hide_const("RailsLiveReload")
-        allow(Rails).to receive(:env).and_return(ActiveSupport::StringInquirer.new("development"))
-        allow(Rails.application).to receive(:root).and_return(described_class.root.join("spec/dummy"))
-      end
-
-      it { is_expected.to be_falsey }
-    end
-  end
-
-  # Two watchers over the same tree fight over one socket, and the gem offers no
-  # hook to skip its own. It is disabled for the length of that one initializer
-  # instead, so both halves have to stay paired.
-  describe "toggling the gem's own watcher" do
-    # The pair hands the previous setting over in an instance variable, so both
-    # halves have to run against the engine instance they are bound to at boot.
-    # The unbound class level collection carries a nil context.
-    def initializer(name)
-      described_class.instance.initializers.detect { _1.name == name }
-    end
-
-    around do |example|
-      was_enabled = RailsLiveReload.config.enabled
-      example.run
-      RailsLiveReload.config.enabled = was_enabled
-    end
-
-    context "when Alchemy is developed from a checkout" do
-      before do
-        allow(described_class).to receive(:watch_engine_files?).and_return(true)
-        RailsLiveReload.config.enabled = true
-      end
-
-      it "disables the gem's watcher before the gem starts it" do
-        initializer("alchemy.disable_live_reload_watcher").run(Rails.application)
-
-        expect(RailsLiveReload.config.enabled).to be(false)
-      end
-
-      it "restores the previous setting once the gem is past it" do
-        initializer("alchemy.disable_live_reload_watcher").run(Rails.application)
-        initializer("alchemy.restore_live_reload").run(Rails.application)
-
-        expect(RailsLiveReload.config.enabled).to be(true)
-      end
-    end
-
-    context "when Alchemy is installed as a gem" do
-      before do
-        allow(described_class).to receive(:watch_engine_files?).and_return(false)
-        RailsLiveReload.config.enabled = true
-      end
-
-      it "leaves the gem's watcher alone" do
-        initializer("alchemy.disable_live_reload_watcher").run(Rails.application)
-
-        expect(RailsLiveReload.config.enabled).to be(true)
-      end
-    end
-
-    it "disables the watcher after the gem's middleware and before its watcher" do
-      expect(initializer("alchemy.disable_live_reload_watcher").after).to eq("rails_live_reload.middleware")
-      expect(initializer("alchemy.disable_live_reload_watcher").before).to eq("rails_live_reload.watcher")
-    end
-
-    it "restores the setting after the gem's watcher and before its metrics" do
-      expect(initializer("alchemy.restore_live_reload").after).to eq("rails_live_reload.watcher")
-      expect(initializer("alchemy.restore_live_reload").before).to eq("rails_live_reload.configure_metrics")
-    end
-  end
-
   describe "alchemy.importmap" do
     let(:additional_importmap) do
       Alchemy::Configurations::Importmap.new(
